refactor: Move HelmChart into own package
