+++ pkg/kluctl_project/target_context.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/kluctl/kluctl/v2/pkg/deployment"
+	"github.com/kluctl/kluctl/v2/pkg/helm"
 	"github.com/kluctl/kluctl/v2/pkg/k8s"
 	"github.com/kluctl/kluctl/v2/pkg/status"
 	"github.com/kluctl/kluctl/v2/pkg/types"
@@ -37,7 +38,7 @@ type TargetContextParams struct {
 	ForSeal            bool
 	Images             *deployment.Images
 	Inclusion          *utils.Inclusion
-	HelmCredentials    deployment.HelmCredentialsProvider
+	HelmCredentials    helm.HelmCredentialsProvider
 	RenderOutputDir    string
 }
